Handling version specific option of bindfs with system without sort -V command available
Some old system did not have -V option avalaible to sort command. This break the version comparaison and did not permit to pass the right parameters to bindfs older than 1.12. So we handle the version comparaison internaly without the use of test and sort command.
This is permit to use vagrant-binds with Red Hat Enterprise Linux Server release 5.10 with bindfs 1.8.3 rpm.