Allow to specify the version of bindfs to be installed (#54)
Extend support of source installation of bindfs to every platform and
add two configuration options:

* bindfs_version
  Specify the version of bindfs to be installed
* install_bindfs_from_source
  Force bindfs to be installed from sources

Unless forced to be installed from source, the required version of
bindfs is first researched in repositories when possible. If no version
match the requirement, fallback to source complation.

This feature only works with exact version match and does not try to
resolve dependencies.