Add a new `force_empty_mountpoints` configuration option
vagrant-bindfs was not used to do anything to mount points before mount,
except ensuring they exist. Hence, if you tried to bind a folder to a
non-empty directory without explicitly allowing Fuse to do so (with the
`nonempty` Fuse option), mount will fail.

This add a new configuration option to the vagrant-bindfs plugin to
make sure mount points are empty. Unless you allow Fuse to not mind with
the `nonempty` Fuse option, mount point will first be destroyed (with
`rm -rf`) then recreated (with `mkdir -p`) before mount.

As this can be dangerous and covers only really specific use cases, it
is disabled by default.