Remove the upper constraint on Ruby version in Gemspec [Fix #92]
Vagrant has its own constraint on Ruby version and it might change
without warning. To remove the upper limit in vagrant-bindfs makes it
easier to deal with Vagrant updates on this constraint without having to
watch it from release to release.

It also allow users with more recent version on Ruby (or pre-release of
Vagrant) to install the plugin.